+++ src/main/java/com/example/goodsprice/config/AsyncConfiguration.java
@@ -6,11 +6,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 @Slf4j
 @Configuration
 @EnableAsync
+@EnableScheduling
 public class AsyncConfiguration {
 
   @Bean(name = "receiptProcessorExecutor")