+++ src/main/java/com/example/goodsprice/config/JacksonConfiguration.java
@@ -1,7 +1,9 @@
 package com.example.goodsprice.config;
 
 import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import java.time.OffsetDateTime;
 import org.openapitools.jackson.nullable.JsonNullableModule;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -17,9 +19,13 @@ public class JacksonConfiguration {
 
   @Bean
   public Jackson2ObjectMapperBuilder objectMapperBuilder() {
+    var offsetDateTimeModule = new SimpleModule();
+    offsetDateTimeModule.addDeserializer(
+        OffsetDateTime.class, new LenientOffsetDateTimeDeserializer());
+
     Jackson2ObjectMapperBuilder builder = new Jackson2ObjectMapperBuilder();
     builder
-        .modules(new JsonNullableModule(), new JavaTimeModule())
+        .modules(new JsonNullableModule(), new JavaTimeModule(), offsetDateTimeModule)
         .featuresToDisable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     return builder;
   }